+++ lib/service/Assistant.dart
@@ -61,5 +61,79 @@ class Assistant /* with Service */ {
     }
   }
 
-  bool get isEnabled => StringUtils.isNotEmpty(Config().wellnessUrl);
+  Future<Message?> sendFeedback(Message message) async {
+    if (!isEnabled) {
+      Log.w('Failed to send assistant feedback. Missing assistant url.');
+      return null;
+    }
+
+    String url = '${Config().assistantUrl}/feedback';
+    String? apiKey = Config().assistantAPIKey;
+    if (apiKey == null) {
+      Log.w('Failed to send assistant feedback. Missing assistant api key.');
+      return null;
+    }
+    Map<String, String> headers = {
+      'Content-Type': 'application/json',
+      'X-API-KEY': apiKey,
+    };
+    Map<String, dynamic> body = {
+      'message_id': message.id,
+      'feedback': message.feedback?.name,
+      'explanation': message.feedbackExplanation,
+    };
+
+    String? json = JsonUtils.encode(body);
+    Response? response = await Network().post(url, auth: Auth2(), headers: headers, body: json, timeout: 30);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      Map<String, dynamic>? responseJson = JsonUtils.decodeMap(responseString);
+      Map<String, dynamic>? answerJson = responseJson?['answer'];
+      if (answerJson != null) {
+        return Message.fromAnswerJson(answerJson);
+      }
+      return null;
+    } else {
+      Log.w('Failed to load assistant feedback response. Response:\n$responseCode: $responseString');
+      return null;
+    }
+  }
+
+
+  Future<int?> getQueryLimit() async {
+    if (!isEnabled) {
+      Log.w('Failed to get assistant query limit. Missing assistant url.');
+      return null;
+    }
+
+    String url = '${Config().assistantUrl}/query-limit';
+    String? apiKey = Config().assistantAPIKey;
+    if (apiKey == null) {
+      Log.w('Failed to get assistant query limit. Missing assistant api key.');
+      return null;
+    }
+    Map<String, String> headers = {
+      'Content-Type': 'application/json',
+      'X-API-KEY': apiKey,
+    };
+
+    Response? response = await Network().get(url, auth: Auth2(), headers: headers);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      Map<String, dynamic>? responseJson = JsonUtils.decodeMap(responseString);
+      dynamic limit = responseJson?['limit'];
+      if (limit is int) {
+        return limit;
+      }
+      return null;
+    } else {
+      Log.w('Failed to load assistant query limit response. Response:\n$responseCode: $responseString');
+      return null;
+    }
+  }
+
+
+  bool get isEnabled => StringUtils.isNotEmpty(Config().assistantUrl);
 }