+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [5.0.19] - 2023-07-04
 ### Added
 - Illinois Assistant query limits and feedback integration [#3349](https://github.com/rokwire/illinois-app/issues/3349)
 - Initialize simplified default content for Favorites panel [#3346](https://github.com/rokwire/illinois-app/issues/3346).